Adjust task due date for bank holidays
Suppliers need to have at least 5 working days to make their
submissions. To now we've simply set the due date to 7 calendar days
into the month to achieve this. This won't be enough come the January
submission window because 1st Jan is a bank holiday. Therefore we need
to dynamically adjust the due date to account for any bank holidays in
the submission period.

Note that this current way to calculate the due date will need adjusting
when we move to modelling submission windows that vary between
frameworks: either 5 working days or 7 calendar days depending on the
specifics of the contract for that framework.
